buggy edit button
bug 1)When deleting MULTIPLE elements in a row, the last element will not delete the first time. BUG 2) The welcome message is bugging, not loading with the rest of the page, loads after. BUG 3) For some reason the divs in the event page all disappear when you click submit button and another button